+++ src/felupe/constitution/autodiff/jax/__init__.py
@@ -1,5 +1,6 @@
+from . import models
 from ._hyperelastic import Hyperelastic
 from ._material import Material
 from ._tools import vmap
 
-__all__ = ["Hyperelastic", "Material", "vmap"]
+__all__ = ["Hyperelastic", "Material", "models", "vmap"]